Misc move/memory tweaks
Disallow max moves from Sketch
Hide dmax moves from legal dropdown lists
Pass ushort for moves for validating memories
Internal class for move pp (hide empty class from dll users)

+++ PKHeX.Core/PKM/PKM.cs
@@ -850,7 +850,7 @@ public abstract class PKM : ISpeciesForm, ITrainerID, IGeneration, IShiny, ILang
     /// </summary>
     /// <param name="move">Move ID</param>
     /// <returns>Amount of PP the move has by default (no PP Ups).</returns>
-    private int GetBasePP(int move) => MoveInfo.GetPP(Context, move);
+    private int GetBasePP(int move) => MoveInfo.GetPP(Context, (ushort)move);
 
     /// <summary>
     /// Applies a shiny <see cref="PID"/> to the <see cref="PKM"/>.
